Move upgrade methods from Request to Response
